backend/requirements.txt
django>=4.2,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
openai>=1.30
google-generativeai>=0.5.2
pydantic>=2.7
python-decouple>=3.8
psycopg2-binary>=2.9
dj-database-url>=2.0
celery>=5.4
django-celery-results>=2.5
pywin32>=306
